feat: implement RichDisplay in UrbanDictionary search
This way it becomes far easier to navigate through pages and since it's
all just 1 API call there really is no reason to not give the option to
go through all pages imo. I also made it so the definitions are properly
sorted by the difference of thumbs up - thumbs down.

Signed-off-by: Jeroen Claassens <support@favware.tech>