feat: Added pnodes, role check, fallback, etc
This is the complete permissions check for whether or not a user can use the `settings` command, which is required for the dashboard to work.

By allowing or denying permissions via Permission Nodes to a user/role, they will not longer be able to access to the dashboard, or rather with this commit, they won't have manageable true.

The API still needs to be updated to do those checks, but will do later. This works perfectly for a MVP.
Fixes #614