fix(weather): wrap long city names (#697)
This makes the attachment overall a bit larger but I tested it against
both a short name (just a city name) and the longest street name in The
Netherlands which is crazy long and fitted so I think this will fit just
about all cases. The prefecture part was moved to be above the city name
so to not have a bit void of space between a short name and the
prefecture.