fix(content, eval): fixes for length
Previously if the content or evalled result was about 1990 characters
(6 backticks, the "md" and the enters) it would throw an error about
message length. To fix this I ported the uploading as file or hastebin
(leaving out logs) from Eval to Content. I also noticed Eval was
checking for a hard 2K characters which also caused issues for the same
reason so dropped it to 1980 as well. While I was working I also added
the last bit of translations for Eval since it was needed for Content
anyway.