refactor(audio): allow no redis
This was previously coded to always try to connect to Redis, even when
Lavalink was disabled. This causes issues in development.

`!.` is safe for events because they are never emitted if Lavalink is
off

`!.` is save for `klasaReady` and `SkyraGuild` because those are not
accessed when Lavalink is off

`?.`  plus ?? is required for Analytics because Influx expects the
function to return points in integer.