fix(readme): partial rewrite of dependencies list (#1656)
The list of dependencies still wasn't really in line with the rest of the changes from #1652 ; that's an oversight on my part sorry.

- Removed "you" usage for lavalink line (and equivalent sentence restructure)
- Removed personification for InfluxDB line
- Corrected period placement to the end of the list

Depending on your choice of grammar a comma could be also added to the end of each list item. For now I haven't added them, though this can evidently be changed based on maintainer preferences.